refactor: move function atbash to a separate module
